+++ Tests/UnityTests.cs
@@ -62,10 +62,19 @@ namespace ObfuscarTests
             TestHelper.CleanInput();
 
             // build it with the keyfile option (embeds the public key, and signs the assembly)
-            File.Copy(Path.Combine(TestHelper.InputPath, @"..", "Microsoft.Practices.Unity.dll"),
-                Path.Combine(TestHelper.InputPath, "Microsoft.Practices.Unity.dll"), true);
-            File.Copy(Path.Combine(TestHelper.InputPath, @"..", "Microsoft.Practices.ServiceLocation.dll"),
-                Path.Combine(TestHelper.InputPath, "Microsoft.Practices.ServiceLocation.dll"), true);
+            string destFileName = Path.Combine(TestHelper.InputPath, "Microsoft.Practices.Unity.dll");
+            if (!File.Exists(destFileName))
+            {
+                File.Copy(Path.Combine(TestHelper.InputPath, @"..", "Microsoft.Practices.Unity.dll"),
+                    destFileName, true);
+            }
+
+            string destFileName1 = Path.Combine(TestHelper.InputPath, "Microsoft.Practices.ServiceLocation.dll");
+            if (!File.Exists(destFileName1))
+            {
+                File.Copy(Path.Combine(TestHelper.InputPath, @"..", "Microsoft.Practices.ServiceLocation.dll"),
+                    destFileName1, true);
+            }
 
             var map = TestHelper.Obfuscate(xml).Mapping;
